% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlRequest.R
\name{as.character.BdlRequestBuilder}
\alias{as.character.BdlRequestBuilder}
\title{Convert a BdlRequestBuilder instance to a character string}
\usage{
\method{as.character}{BdlRequestBuilder}(x, ...)
}
\arguments{
\item{x}{A BdlRequestBuilder object}

\item{...}{Any additional parameters}
}
\description{
Convert a BdlRequestBuilder instance to a character string
}

