% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BdlConnection.R
\name{BdlConnection}
\alias{BdlConnection}
\title{Create an FTP connection to Bloomberg Datalicense}
\usage{
BdlConnection(user, pw, key)
}
\arguments{
\item{user}{The account number assigned by Bloomberg}

\item{pw}{The password of your Bloomberg account}

\item{key}{The DES decryption key of your Bloomberg account}
}
\value{
an S3 BdlConnection object, used to upload requests and download prices.
}
\description{
Create an FTP connection to Bloomberg Datalicense
}
\examples{
#these are dummy credentials. Replace with the credentials received from Bloomberg
con <- BdlConnection(user = 'dl111111',
                     pw = 'XvH,gE2A',
                     key = '3xzZl0yA')
}

