% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlRequest.R
\name{BdlRequest}
\alias{BdlRequest}
\title{Create a Datalicense request}
\usage{
BdlRequest(firmName, programName, fields, instruments, header = c())
}
\arguments{
\item{firmName}{The firm name assigned to you by Bloomberg}

\item{programName}{The name of the Datalicense program. Namely:
getdata, gethistory, getquotes, getallquotes, getactions, getcompany, getsnap}

\item{fields}{A vector containing the fields to be downloaded,
e.g. fields = c('PX_LAST', 'PX_CLOSE')}

\item{instruments}{A vector containing th instruments to be downloaded,
e.g. instruments = c('IBM US Equity')}

\item{header}{A vector containing additional header fields,
e.g. header = c(PROGRAMFLAG = 'oneshot', SECMASTER = 'yes')}
}
\value{
a BdlRequest object
}
\description{
Create a Datalicense request
}

