% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BdlRequest.R
\name{BdlRequestBuilder}
\alias{BdlRequestBuilder}
\title{Helper to create a Datalicense request}
\usage{
BdlRequestBuilder(header, fields, data)
}
\arguments{
\item{header}{A vector containing additional header fields, 
e.g. header = c(FIRMNAME = 'dl1234', PROGRAMNAME = 'getdata', PROGRAMFLAG = 'oneshot', SECMASTER = 'yes')}

\item{fields}{A vector containing the fields to be downloaded, 
e.g. fields = c('PX_LAST', 'PX_CLOSE')}

\item{data}{A vector containing the instruments to be downloaded, 
e.g. data = c('IBM US Equity')}
}
\value{
a BdlRequestBuilder object
}
\description{
Please refer to the Bloomberg Per Security Enterprise Solution manual for details.
}

