% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlRequest.R
\name{BdlRequestGetData}
\alias{BdlRequestGetData}
\title{Create a GetData request}
\usage{
BdlRequestGetData(firmName, fields, instruments, header = c())
}
\value{
a BdlRequest object
}
\description{
Create a GetData request
}

