% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlResponse.R
\name{BdlResponse}
\alias{BdlResponse}
\title{Handle to let you retrieve a data file from Bloomberg}
\usage{
BdlResponse(bdlRequest, bdlConnection, targetFileName)
}
\arguments{
\item{bdlRequest}{The request}

\item{bdlConnection}{The connection}

\item{targetFileName}{The name of the file at Bloomberg}
}
\description{
Typically, this is not called directly, but returned when you call the
}
\seealso{
UploadRequest

BdlRequest
}

