% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libdes.R
\docType{package}
\name{data.tree}
\alias{data.tree}
\alias{data.tree-package}
\title{Encrypt and decrypt files and R objects with the DES algorithm}
\description{
Encrypt and decrypt files and R objects with the DES algorithm
}

