% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datalicenseR.R
\docType{package}
\name{datalicenseR}
\alias{datalicenseR}
\alias{datalicenseR-package}
\title{Download market data from Bloomberg Datalicense}
\description{
\code{datalicenseR} lets you download price data directly from Bloomberg. Prerequisite is a valid Datalicense with Bloomberg.
The main high level functions you can use to quickly download and parse data from Bloomberg are:
\describe{
  \item{\code{\link{GetData}}}{To use the bloomberg getdata program.}
  \item{\code{\link{GetHistory}}}{To use the bloomberg gethistory program.}
  \item{\code{\link{GetSnapshot}}}{To use the bloomberg getsnap program.}
}
}
\seealso{
For more details and lower level functions, see the \code{datalicenseR} vignette by running: \code{vignette("datalicenseR")}
}

