% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libdes_functions.R
\name{Decrypt}
\alias{Decrypt}
\title{Decrypt a character string}
\usage{
Decrypt(text, key)
}
\arguments{
\item{text}{An encrypted character string}

\item{key}{The key as a character string}
}
\description{
Decrypt a character string
}

