% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libdes_functions.R
\name{DecryptFile}
\alias{DecryptFile}
\title{Decrypt a file with the DES algorithm}
\usage{
DecryptFile(sourceFile, decryptedFile, key, SUNOS = FALSE, HEX_KEY = FALSE,
  ECB = FALSE, UUENC = FALSE)
}
\arguments{
\item{sourceFile}{The path to the file that should be decrypted}

\item{decryptedFile}{The path to the target file}

\item{key}{The key (either a character string of arbitrary length, or a 56bit HEX key)}

\item{SUNOS}{Use SUNOS?}

\item{HEX_KEY}{Should the key be interpreted as HEX?}

\item{ECB}{Use ECB?}

\item{UUENC}{Whether to use UU Encoding}
}
\description{
Decrypt a file with the DES algorithm
}

