% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libdes_functions.R
\name{DecryptFileTripleDES}
\alias{DecryptFileTripleDES}
\title{Decrypt a file with the Triple DES algorithm}
\usage{
DecryptFileTripleDES(sourceFile, decryptedFile, key, ECB = FALSE,
  UUENC = FALSE)
}
\arguments{
\item{sourceFile}{The path the the file that should be decrypted}

\item{decryptedFile}{The path to the target file}

\item{key}{The 56bit HEX key}

\item{ECB}{Use ECB?}

\item{UUENC}{Whether the files should be UU encoded}
}
\description{
Decrypt a file with the Triple DES algorithm
}

