% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BdlResponse.R
\name{DeriveResponseFileName}
\alias{DeriveResponseFileName}
\title{Derives the expected response file name}
\usage{
DeriveResponseFileName(bdlRequest = NULL, requestFileName = NULL,
  responseFileName = NULL)
}
\arguments{
\item{bdlRequest}{A BdlRequestBuilder object, or a character string 
or a character string containing the request content, optional}

\item{requestFileName}{The target file name of the request at the Bloomberg FTP site, optional}

\item{responseFileName}{The name of the response file. If omitted, the method deducts 
the response file name either from the BdlRequestBuilder, or from the requestFileName}
}
\value{
a character string, representing the responseFileName derived from the arguments provided
}
\description{
Derives the expected response file name
}
\seealso{
BdlRequestBuilder
}

