% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BdlConnection.R
\name{DownloadFTP_2}
\alias{DownloadFTP_2}
\title{Downloads a file via FTP.}
\usage{
DownloadFTP_2(baseURL, filePath, delete = FALSE)
}
\arguments{
\item{baseURL}{The base URL}

\item{filePath}{The path to the file, relative from the baseURL}

\item{delete}{Whether or not to delete the file after successful downloading}
}
\value{
An FTPDownloadResult
}
\description{
Downloads a file via FTP.
}
\seealso{
FTPDownloadResult
}

