% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libdes_functions.R
\name{Encrypt}
\alias{Encrypt}
\title{Encrypt a character string}
\usage{
Encrypt(text, key)
}
\arguments{
\item{text}{A character string}

\item{key}{The key as a character string}
}
\description{
Encrypt a character string
}

