% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libdes_functions.R
\name{EncryptFile}
\alias{EncryptFile}
\title{Encrypt a file with the DES algorithm}
\usage{
EncryptFile(sourceFile, encryptedFile, key, SUNOS = FALSE, HEX_KEY = FALSE,
  ECB = FALSE, UUENC = FALSE, uuencFileName = "")
}
\arguments{
\item{sourceFile}{The path to the file that should be encrypted}

\item{encryptedFile}{The path to the target file}

\item{key}{The key (either a character string of arbitrary length, or a 56bit HEX key)}

\item{SUNOS}{Use SUNOS?}

\item{HEX_KEY}{Should the key be interpreted as HEX?}

\item{ECB}{Use ECB?}

\item{UUENC}{Whether to use UU Encoding}

\item{uuencFileName}{The original file name to add as a header to the Uuencoded file}
}
\description{
Encrypt a file with the DES algorithm
}

