% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HighLevelFunctions.R
\name{GetData}
\alias{GetData}
\title{Download data from Bloomberg using the getdata program.}
\usage{
GetData(user, pw, key, fields, tickers, parser = GetDataParser)
}
\arguments{
\item{user}{The account number assigned by Bloomberg}

\item{pw}{The password of your Bloomberg account}

\item{key}{The DES decryption key of your Bloomberg account}

\item{fields}{A vector of fields, e.g. c('PX_LAST', 'NAME')}

\item{tickers}{A vector of Bloomberg tickers to be downloaded, e.g. c('SPX Index', 'IBM US Equity')}

\item{parser}{The parser used to convert the file into an R object}
}
\value{
An R object containing the downloaded data. If you use the default parser, then a data.frame
is returned.
}
\description{
This methods blocks call, until the file is available (which can be several minutes)
}

