% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ErrorCodes.R
\name{GetDataErrorMessage}
\alias{GetDataErrorMessage}
\title{Convert an Error Code from a GetData request to an Error Message}
\usage{
GetDataErrorMessage(errorCode)
}
\arguments{
\item{errorCode}{The Bloomberg Error Code}
}
\value{
a human readable error message
}
\description{
Convert an Error Code from a GetData request to an Error Message
}

