% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{getDataReply}
\alias{getDataReply}
\title{getDataReply}
\format{A data.frame}
\usage{
data(getDataReply)
}
\description{
A sample reply having used the \code{\link{GetDataParser}}.
}
\details{
\itemize{
  \item ERROR_CODE 0 if ok
  \item PX_LAST the value of the PX_LAST field
  \item PX_CLOSE the value of the PX_CLOSE field
}
}
\keyword{datasets}

