% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ErrorCodes.R
\name{GetHistoryErrorMessage}
\alias{GetHistoryErrorMessage}
\title{Convert an Error Code from a GetHistory request to an Error Message}
\usage{
GetHistoryErrorMessage(errorCode)
}
\arguments{
\item{errorCode}{The Bloomberg Error Code}
}
\value{
a human readable error message
}
\description{
Convert an Error Code from a GetHistory request to an Error Message
}

