% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{GetHistoryListParser}
\alias{GetHistoryListParser}
\title{Convert Bloomberg gethistory out file content to a list of xts objects}
\usage{
GetHistoryListParser(bdlOutContent)
}
\arguments{
\item{bdlOutContent}{The content string}
}
\value{
A list containing the xts objects
}
\description{
Convert Bloomberg gethistory out file content to a list of xts objects
}
\examples{
#to look at an example:
data(getHistoryListReply)
getHistoryListReply
}
\seealso{
GetHistoryParser
}

