% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{getHistoryListReply}
\alias{getHistoryListReply}
\title{getHistoryListReply}
\format{A \code{\link{list}} containing \code{\link{xts}} objects.}
\usage{
data(getHistoryListReply)
}
\description{
A sample reply having used the \code{\link{GetHistoryListParser}}.
}
\keyword{datasets}

