% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{GetHistoryParserList}
\alias{GetHistoryParserList}
\title{Convert Bloomberg out file content to a list of xts objects}
\usage{
GetHistoryParserList(bdlOutContent)
}
\arguments{
\item{bdlOutContent}{The content string}
}
\value{
A list containing the xts objects
}
\description{
Convert Bloomberg out file content to a list of xts objects
}

