% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{getHistoryReply}
\alias{getHistoryReply}
\title{getHistoryReply}
\format{An \code{\link{xts}} object}
\usage{
data(getHistoryReply)
}
\description{
A sample reply having used the \code{\link{GetHistoryParser}}.
}
\keyword{datasets}

