% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorCodes.R
\name{GetSnapshotErrorMessage}
\alias{GetSnapshotErrorMessage}
\title{Convert an Error Code from a Snapshot to an Error Message}
\usage{
GetSnapshotErrorMessage(errorCode)
}
\arguments{
\item{errorCode}{The Bloomberg Error Code}
}
\value{
a human readable error message
}
\description{
Convert an Error Code from a Snapshot to an Error Message
}

