% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{getSnapshotReply}
\alias{getSnapshotReply}
\title{getSnapshotReply}
\format{A \code{\link{data.frame}} containing the reply data.}
\usage{
data(getSnapshotReply)
}
\description{
A sample reply having used the \code{\link{GetSnapshotReplyParser}}.
}
\keyword{datasets}

