% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{GetSnapshotReplyParser}
\alias{GetSnapshotReplyParser}
\title{Convert Bloomberg out file content to data.frame}
\usage{
GetSnapshotReplyParser(bdlOutContent)
}
\arguments{
\item{bdlOutContent}{The content string}
}
\value{
an data frame object containing the price series
}
\description{
Convert Bloomberg out file content to data.frame
}

