% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{getSnapshotResponse}
\alias{getSnapshotResponse}
\title{getSnapshotResponse}
\format{A \code{\link{data.frame}} containing the response data.}
\usage{
data(getSnapshotResponse)
}
\description{
A sample response having used the \code{\link{GetSnapshotResponseParser}}.
}
\keyword{datasets}

