% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{GetSnapshotResponseParser}
\alias{GetSnapshotResponseParser}
\title{Convert Bloomberg out file content to data.frame}
\usage{
GetSnapshotResponseParser(bdlOutContent)
}
\arguments{
\item{bdlOutContent}{The content string}
}
\value{
a data frame object containing the price series
}
\description{
The Response file can be used to check if there are any exceptions.
It does not contain the actual price data.
}

