% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlResponse.R
\name{ParseBdlResponse}
\alias{ParseBdlResponse}
\title{Convert Bloomberg out file content to data.frame}
\usage{
ParseBdlResponse(bdlOutContent)
}
\arguments{
\item{bdlOutContent}{The content string}
}
\value{
a data.frame
}
\description{
Convert Bloomberg out file content to data.frame
}

