% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlRequest.R
\name{print.BdlRequestGetData}
\alias{print.BdlRequestGetData}
\title{Convert a BdlRequestGetData instance to a character string}
\usage{
\method{print}{BdlRequestGetData}(bdlRequest, ...)
}
\description{
Convert a BdlRequestGetData instance to a character string
}

