% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{response1}
\alias{response1}
\title{response1: A sample response}
\format{A data.frame}
\usage{
data(response1)
}
\description{
The response received when uploading tests/data/sample1.req
}
\details{
\itemize{
  \item Status 0 if ok
  \item PX_LAST the value of the PX_LAST field
  \item PX_CLOSE the value of the PX_CLOSE field
}
}
\keyword{datasets}

