% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlResponse.R
\name{TryGetBdlData}
\alias{TryGetBdlData}
\title{Download the data, if available}
\usage{
TryGetBdlData(bdlConnection, responseFileName)
}
\arguments{
\item{bdlConnection}{The BdlConnection object used to establish the FTP download}

\item{responseFileName}{The file downloaded}
}
\value{
either NULL (if the file is not yet available) or a data.frame containing the data
}
\description{
Download the data, if available
}
\seealso{
UploadRequest

BdlResponseHandle
}

