% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlConnection.R
\name{UploadRequest}
\alias{UploadRequest}
\title{Uploads a BdlRequest to Bloomberg}
\usage{
UploadRequest(bdlRequest, bdlConnection, targetFilePath)
}
\arguments{
\item{bdlRequest}{A BdlRequest object}

\item{bdlConnection}{A BdlConnection object}

\item{targetFilePath}{The name the file should have at the Bloomberg FTP site}
}
\value{
BdlResponse
}
\description{
Uploads a BdlRequest to Bloomberg
}
\seealso{
BdlConnection

BdlRequest

BdlResponse
}

