% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BdlConnection.R
\name{BdlConnection}
\alias{BdlConnection}
\title{Create an FTP connection to Bloomberg Datalicense}
\usage{
BdlConnection(user, pw, key, sftp = FALSE)
}
\arguments{
\item{user}{The account number assigned by Bloomberg}

\item{pw}{The password of your Bloomberg account}

\item{key}{The DES decryption key of your Bloomberg account}

\item{sftp}{if TRUE, then an sftp connection is established. On some Linux systems, this may
not work out of the box, as libcurl does not natively support sftp. In that case, you need to 
compile curl with SFTP support. See here for details: http://askubuntu.com/questions/195545/how-to-enable-sftp-support-in-curl}
}
\value{
an S3 BdlConnection object, used to upload requests and download prices.
}
\description{
Create an FTP connection to Bloomberg Datalicense
}
\examples{
#these are dummy credentials. Replace with the credentials received from Bloomberg 
con <- BdlConnection(user = 'dl111111', 
                     pw = 'XvH,gE2A', 
                     key = '3xzZl0yA')

}

