% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlConnection.R
\name{BdlConnection}
\alias{BdlConnection}
\title{Create an FTP or SFTP connection to Bloomberg Datalicense}
\usage{
BdlConnection(user, pw, key)
}
\arguments{
\item{user}{The account number assigned by Bloomberg}

\item{pw}{The password of your Bloomberg account}

\item{key}{The DES decryption key of your Bloomberg account}
}
\value{
an S3 BdlConnection object, used to upload requests and download prices.
}
\description{
Create an FTP or SFTP connection to Bloomberg Datalicense
}

