% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{BdlConnection}
\alias{BdlConnection}
\title{Create an FTP or SFTP connection to Bloomberg Datalicense}
\usage{
BdlConnection(account, pw, host = "bfmrr.bloomberg.com",
  ftpType = "passive", encrypt = FALSE, port = 21,
  passivePorts = c(32768, 65535))
}
\arguments{
\item{account}{The account number assigned by Bloomberg}

\item{pw}{The password of your Bloomberg account}

\item{host}{The Bloomberg ftp host}

\item{ftpType}{Any of \code{passive}, \code{active}}

\item{encrypt}{If \code{TRUE}, then SFTP is used}

\item{port}{The port used by FTP (20 for active, 21 for passive connections) or by SFTP (30206)}

\item{passivePorts}{In the case of ftpMode = 'passive', this is the min and the max port for the data port range.}
}
\value{
an S3 BdlConnection object, used to upload requests and download prices.
}
\description{
Create an FTP or SFTP connection to Bloomberg Datalicense
}

