% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datalicenseR.R
\docType{package}
\name{datalicenseR}
\alias{datalicenseR}
\alias{datalicenseR-package}
\title{Download market data from Bloomberg Datalicense}
\description{
\code{datalicenseR} lets you download price data directly from Bloomberg. Prerequisite is a valid Datalicense with Bloomberg.
}
\seealso{
For more details see the \code{datalicenseR} vignette by running: \code{vignette("datalicenseR")}
}

