% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlResponse.R
\name{DownloadResponse}
\alias{DownloadResponse}
\title{Download the data in sync mode, waiting until the result file is there}
\usage{
DownloadResponse(bdlConnection, responseFileName)
}
\arguments{
\item{bdlConnection}{The BdlConnection object used to establish the FTP download}

\item{responseFileName}{The file downloaded}
}
\value{
the response content
}
\description{
Download the data in sync mode, waiting until the result file is there
}
\seealso{
UploadRequest

BdlResponseHandle
}

