% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{GetDataParser}
\alias{GetDataParser}
\title{Convert Bloomberg out file content to data.frame}
\usage{
GetDataParser(bdlOutContent, columnHeader = TRUE)
}
\arguments{
\item{bdlOutContent}{The content string}

\item{columnHeader}{If TRUE, then column headers are auto detected. If FALSE, then it
is assumed no column headers exist.}
}
\value{
a data.frame
}
\description{
Convert Bloomberg out file content to data.frame
}

