% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HighLevelFunctions.R
\name{GetHistory}
\alias{GetHistory}
\title{Download data from Bloomberg using the gethistory program.}
\usage{
GetHistory(user, pw, key, fields, tickers, fromDate, toDate = Sys.Date(),
  sync = TRUE, parser = GetHistoryParser, verbose = FALSE)
}
\arguments{
\item{user}{The account number assigned by Bloomberg}

\item{pw}{The password of your Bloomberg account}

\item{key}{The DES decryption key of your Bloomberg account}

\item{fields}{A vector of fields, e.g. c('PX_LAST', 'NAME')}

\item{tickers}{A vector of Bloomberg tickers to be downloaded, e.g. c('SPX Index', 'IBM US Equity')}

\item{fromDate}{The start date in your request}

\item{toDate}{The end date in your request}

\item{sync}{If TRUE, then the call will not return until the file is available, or until it times out}

\item{parser}{The parser used to convert the file into an R object}

\item{verbose}{Prints output if TRUE}
}
\value{
A list containing the downloaded data. If you use the default parser, then an xts object
is returned.
}
\description{
This methods blocks call, until the file is available (which can be several minutes)
}

