% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{GetHistoryParser}
\alias{GetHistoryParser}
\title{Convert Bloomberg out file content to data.frame}
\usage{
GetHistoryParser(bdlOutContent)
}
\arguments{
\item{bdlOutContent}{The content string}
}
\value{
an xts object containing the price series
}
\description{
Convert Bloomberg out file content to data.frame
}

