% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\name{GetHistoryParser}
\alias{GetHistoryParser}
\title{Convert Bloomberg out file content to a single xts}
\usage{
GetHistoryParser(bdlOutContent)
}
\arguments{
\item{bdlOutContent}{The content string}
}
\value{
an xts object containing the price series
}
\description{
Convert Bloomberg out file content to a single xts
}
\examples{
#to look at an example:
data(getHistoryReply)
getHistoryReply
 
}
\seealso{
GetHistoryListParser
}

