% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\name{GetSnapshotResponseParser}
\alias{GetSnapshotResponseParser}
\title{Convert Bloomberg resp file content to data.frame}
\usage{
GetSnapshotResponseParser(bdlOutContent)
}
\arguments{
\item{bdlOutContent}{The content string}
}
\value{
a data frame object containing the price series
}
\description{
The Response file can be used to check if there are any exceptions.
It does not contain the actual price data.
}
\examples{
#to look at an example:
data(getSnapshotResponse)
getSnapshotResponse

}

