% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlResponse.R
\name{ParseBdlResponseFile}
\alias{ParseBdlResponseFile}
\title{Convert Bloomberg out file content to data.frame}
\usage{
ParseBdlResponseFile(bdlOutContent)
}
\arguments{
\item{bdlOutContent}{The content string}
}
\value{
a data.frame
}
\description{
Convert Bloomberg out file content to data.frame
}

