% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BdlRequest.R
\name{print.BdlRequest}
\alias{print.BdlRequest}
\title{Convert a BdlRequest instance to a character string}
\usage{
\method{print}{BdlRequest}(x, ...)
}
\arguments{
\item{x}{A BdlRequest}

\item{...}{Any additional parameters}
}
\description{
Convert a BdlRequest instance to a character string
}

