% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BdlResponse.R
\name{TryGetBdlData}
\alias{TryGetBdlData}
\title{Download the data, if available}
\usage{
TryGetBdlData(bdlConnection, responseFileName, parser, verbose = FALSE,
  useSystemUudecode = FALSE)
}
\arguments{
\item{bdlConnection}{The BdlConnection object used to establish the FTP download}

\item{responseFileName}{The file downloaded}

\item{parser}{The parser used to convert the file into an R object}

\item{verbose}{Prints progress output if TRUE}

\item{useSystemUudecode}{workaround for a bug with long filenames on some linux systems. Requires installation
of sharutils on your system (sudo apt-get install sharutils)}
}
\value{
either NULL (if the file is not yet available) or a data.frame containing the data
}
\description{
Download the data, if available
}
\seealso{
UploadRequest

BdlResponseHandle
}

