% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BdlConnection.R
\name{UploadRequest}
\alias{UploadRequest}
\title{Uploads a request file to Bloomberg}
\usage{
UploadRequest(bdlConnection, bdlRequest, targetFileName,
  responseFileName = NULL, verbose = FALSE)
}
\arguments{
\item{bdlConnection}{A BdlConnection object}

\item{bdlRequest}{A BdlRequestBuilder object, or a character string containing the request content}

\item{targetFileName}{The target file name of the request at the Bloomberg FTP site}

\item{responseFileName}{The name of the response file. If omitted, the method deducts
the response file name either from the BdlRequestBuilder, or from the targetFileName}

\item{verbose}{Prints output if TRUE}
}
\value{
A character string, representing the name of the response file
}
\description{
Uploads a request file to Bloomberg
}
\seealso{
BdlConnection

BdlRequestBuilder
}

